windows-build part2

+++ src/Modules/Physics/Components/PhysicsCharacterComponent.cpp
@@ -78,7 +78,7 @@ void PhysicsCharacterComponent::create(TransformComponent &transform) {
     // m_physicsBody->SetRestitution(0);
 
     auto *userData = new PhysicsBodyUserData(m_EntityId.id());
-    m_physicsBody->SetUserData(reinterpret_cast<unsigned long>(userData));
+    m_physicsBody->SetUserData(reinterpret_cast<unsigned long long>(userData));
 
     m_PhysicsResource().getInterface().AddBody(m_physicsBody->GetID(), JPH::EActivation::Activate);
 }
